Reduce protocol to plain strings & implement client
First, the server sends to the client plain strings only.
Second, the textual format of the client input will hold all data necessary to determine what should be done and the server decides if the message is valid or not anyway.
Thus it makes no sense to send anything else over the wire other than plain strings.

This commit also implements the client. Given this very simple protocol the client code does not need to know anything about either the client or the server message format – the user will react to the prompts sent from the server accordingly and, if not, the server will reject the messages anyway. The whole "business" logic is on the server.